+++ R/jsFunc-visibilityFuncs.R
@@ -22,6 +22,9 @@
 #'   \strong{\code{time}}       \tab The number of seconds to make the
 #'                                   animation last
 #'                                   (default: \code{0.5}) \cr
+#'   \strong{\code{delay}}      \tab The number of seconds to wait before
+#'                                   hiding/showing the element
+#'                                   (default: \code{0}) \cr
 #'   \strong{\code{condition}}  \tab An optional argument to \code{toggle}, see
 #'                                   'Details' below. \cr
 #' }
@@ -52,6 +55,7 @@
 #'   toggle(id = "element")
 #'   toggle("element", TRUE)
 #'   toggle("element", TRUE, "fade", 2)
+#'   toggle(id = "element", delay = 1)
 #'   toggle(id = "element", time = 1, anim = TRUE, animType = "slide")
 #'   show("element")
 #'   show(id = "element", anim = TRUE)