+++ inst/examples/sandbox/helpers.R
@@ -1,6 +1,7 @@
 # some sample R expressions to try
 examples <- c(
   'toggle("test")',
+  'toggle(id = "test", delay = 0.5)',
   'toggle("test", TRUE, "fade", 2)',
   'toggle(id = "test", anim = TRUE, time = 2, animType = "slide")',
   'hide("test")',